requirement.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
defusedxml==0.7.1
Django==3.2.13
django-cors-headers==3.12.0
django-filter==21.1
django-guardian==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
httpie==3.2.1
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.7
multidict==6.0.2
mysqlclient==2.1.0
passlib==1.7.4
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
shortuuid==1.0.9
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
uuid==1.30
zipp==3.8.0
